Update to .NET 6.0
