+++ buildspec.yml
@@ -2,7 +2,7 @@ version: 0.2
 phases:
   install:
     runtime-versions:
-        dotnet: 3.1
+        dotnet: 6.0
   pre_build:
     commands:
       - echo Restore started on `date`
